perf(deps): let a build omit the OTLP exporters it does not use
This is the tag that actually releases google.golang.org/grpc. The OTLP
exporters are what pin it in a GoFr binary -- otlptracegrpc from
otel.go, otlpmetricgrpc from metrics/exporters, and otlpmetrichttp too,
which imports grpc for its status codes -- so gofr_nogrpc and
gofr_nodgraph each removed their own package and left the 66-package
gRPC tree standing. With gofr_nootlp set, all three go and the tree goes
with them.

Measured on a minimal GoFr service (gofr.New, one route, Run):

    default                                57.10 MB   831 pkgs   86 grpc
    +nosqldrivers,nopubsub                 42.07 MB   574 pkgs   70 grpc
    +nographql                             41.26 MB   554 pkgs   70 grpc
    +nogrpc                                40.59 MB   543 pkgs   66 grpc
    +nodgraph                              39.47 MB   541 pkgs   66 grpc
    +nootlp                                35.35 MB   414 pkgs    0 grpc

4.12 MB and 127 packages for this tag alone, and 57.10 -> 35.35 MB
(-38.1%) and 831 -> 414 packages (-50.2%) for the six together. The
tags are conjunctive on that tree, not additive: this one is worth 127
packages only because the other two went first.

What the tag removes is a transport, not observability. Prometheus is
untouched and is still the default metrics exporter; gcp brings its own
transport; TRACE_EXPORTER=zipkin and gofr do not go through OTLP and
still export. What a tagged build cannot do is speak OTLP, and it says
so: "otlp" stays registered in the metrics registry and fails at the
point it would open the connection, rather than vanishing and letting
METRICS_EXPORTER=otlp quietly resolve to a different exporter than the
operator configured. Both errors name the tag and the alternatives.

Two defects came out of the code this touches, both pre-existing and
both on the path the tag exercises:

initTracer built a BatchSpanProcessor over whatever getExporter
returned, without checking for nil. getExporter returns a nil exporter
and a nil error for an unsupported TRACE_EXPORTER -- a typo in config is
enough -- and a nil exporter with an error whenever an exporter fails to
build. The processor panics on its first flush, on its own goroutine.
It now returns instead, leaving the tracer provider installed so trace
and span IDs still appear on every request.

initTracer reported exporter failures with container.Error(err), which
JSON-marshals the error value; an errors.errorString has no exported
fields, so every such failure logged {"message":{}} and the reason was
lost. It uses Errorf now. The same pattern is likely elsewhere and is
worth a sweep of its own.
