+++ .github/workflows/go.yml
@@ -568,6 +568,18 @@ jobs:
           go vet   -tags gofr_nodgraph ./pkg/gofr/migration/
           go test  -tags gofr_nodgraph -count=1 ./pkg/gofr/migration/
 
+      # gofr_nootlp is the tag that actually releases google.golang.org/grpc: the
+      # OTLP exporters pin it, otlpmetrichttp included, so gofr_nogrpc and
+      # gofr_nodgraph only pay off once this one is set too. Both halves of the
+      # trace and metric transports have tests that exist only under the tag.
+      - name: Build and test with the OTLP exporters omitted
+        run: |
+          set -euo pipefail
+          go build -tags gofr_nootlp ./...
+          go vet   -tags gofr_nootlp ./pkg/gofr/ ./pkg/gofr/metrics/exporters/
+          go test  -tags gofr_nootlp -count=1 ./pkg/gofr/metrics/exporters/
+          go test  -tags gofr_nootlp -count=1 -run 'TestOTLPTracesOmitted|Test_initTracer' ./pkg/gofr/
+
       # The tags are opt-in, so the default build must stay exactly as it was.
       # This is the assertion that keeps that true.
       - name: Default build links the same packages as before
