+++ pkg/gofr/metrics/exporters/otlp_test.go
@@ -38,40 +38,6 @@ func Test_temporalitySelector(t *testing.T) {
 	}
 }
 
-func Test_buildOTLPExporter(t *testing.T) {
-	tests := []struct {
-		name string
-		cfg  Config
-	}{
-		{"grpc default", Config{Endpoint: "localhost:4317", Protocol: "grpc", Insecure: true}},
-		{"grpc with headers", Config{Endpoint: "localhost:4317", Protocol: "grpc", Insecure: true, Headers: map[string]string{"x": "y"}}},
-		{"grpc secure with delta", Config{Endpoint: "otlp.nr-data.net:4317", Protocol: "grpc", Temporality: "delta"}},
-		{"http host:port", Config{Endpoint: "localhost:4318", Protocol: "http", Insecure: true}},
-		{"http full url", Config{Endpoint: "https://collector.example.com/v1/metrics", Protocol: "http"}},
-		{"grpc https scheme ignores insecure override", Config{
-			Endpoint: "https://collector.example.com:4317", Protocol: "grpc", Insecure: true,
-		}},
-		{"http https scheme ignores insecure override", Config{
-			Endpoint: "https://collector.example.com/v1/metrics", Protocol: "http", Insecure: true,
-		}},
-	}
-
-	for _, tc := range tests {
-		t.Run(tc.name, func(t *testing.T) {
-			exp, err := buildOTLPExporter(context.Background(), &tc.cfg)
-			if err != nil {
-				t.Fatalf("unexpected error: %v", err)
-			}
-
-			if exp == nil {
-				t.Fatal("expected non-nil exporter")
-			}
-
-			_ = exp.Shutdown(context.Background())
-		})
-	}
-}
-
 func Test_buildOTLPReader_emptyEndpoint(t *testing.T) {
 	cfg := Config{Endpoint: "", Protocol: "grpc"}
 
@@ -101,31 +67,5 @@ func Test_buildOTLPReader_pushReaderDegradesOnEmptyEndpoint(t *testing.T) {
 
 	if !strings.Contains(out, "METRICS_URL") {
 		t.Errorf("expected the degrade error to mention METRICS_URL, got: %q", out)
-	}
-}
-
-func Test_buildOTLPReader_warnsOnUnrecognizedProtocolAndTemporality(t *testing.T) {
-	tests := []struct {
-		name    string
-		cfg     Config
-		wantLog string
-	}{
-		{"unrecognized protocol", Config{Endpoint: "localhost:4317", Protocol: "carrier-pigeon"}, "METRICS_PROTOCOL"},
-		{"unrecognized temporality", Config{Endpoint: "localhost:4317", Temporality: "quantum"}, "METRICS_TEMPORALITY"},
-	}
-
-	for _, tc := range tests {
-		t.Run(tc.name, func(t *testing.T) {
-			out := testutil.StdoutOutputForFunc(func() {
-				_, err := buildOTLPReader(context.Background(), &tc.cfg, logging.NewMockLogger(logging.WARN))
-				if err != nil {
-					t.Fatalf("unexpected error: %v", err)
-				}
-			})
-
-			if !strings.Contains(out, tc.wantLog) {
-				t.Errorf("expected warning to mention %q, got: %q", tc.wantLog, out)
-			}
-		})
 	}
 }
